reexpose net classes, rename net_util
