+++ slm_lab/agent/net/__init__.py
@@ -5,6 +5,3 @@ Contains classes of neural network architectures
 
 from slm_lab.agent.net.convnet import ConvNet
 from slm_lab.agent.net.feedforward import MLPNet
-
-nets = {'Conv': ConvNet,
-        'MLP': MLPNet}
