+++ slm_lab/agent/net/net_util.py
@@ -1,0 +1,16 @@
+import torch
+
+
+def flatten_params(net):
+    '''Source: https://discuss.pytorch.org/t/running-average-of-parameters/902/2'''
+    return torch.cat([param.data.view(-1) for param in net.parameters()], 0)
+
+
+def load_params(net, flattened):
+    '''Source: https://discuss.pytorch.org/t/running-average-of-parameters/902/2'''
+    offset = 0
+    for param in net.parameters():
+        param.data.copy_(
+            flattened[offset:offset + param.nelement()]).view(param.size())
+        offset += param.nelement()
+    return net
